Move tests and generated sources to 'standard'-package
