+++ .gitignore
@@ -19,6 +19,9 @@
 *.tar.gz
 *.rar
 
+# SHA 256
+*.sha256
+
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 replay_pid*
