--- src/test/java/com/chrimle/example/TestExampleRecordWithDefaultFields.java
@@ -1,48 +1,0 @@
-package com.chrimle.example;
-
-import com.chrimle.example.utils.AssertionUtils;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-@DisplayName("Testing Record with default fields")
-public class TestExampleRecordWithDefaultFields {
-
-  private final Class<?> classUnderTest = ExampleRecordWithDefaultFields.class;
-
-  @Test
-  @DisplayName("Testing class generated is a Record")
-  public void testClassGeneratedIsRecord() {
-    AssertionUtils.assertIsRecord(classUnderTest);
-  }
-
-  @Test
-  @DisplayName("Testing generated Record is not annotated as @Deprecated")
-  public void testRecordIsNotAnnotatedAsDeprecated() {
-    AssertionUtils.assertClassIsNotAnnotatedAsDeprecated(classUnderTest);
-  }
-
-  @Test
-  @DisplayName("Testing nullable field is null after instantiation")
-  public void testNullableFieldIsNullAfterInstantiation() {
-    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
-        null, "ignore");
-    Assertions.assertNull(record.nullableStringField());
-  }
-
-  @Test
-  @DisplayName("Testing field with default value is set to default if null is provided")
-  public void testFieldWithDefaultValueIsSetToDefaultIfProvidedValueIsNull() {
-    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
-        "ignore", null);
-    Assertions.assertEquals("someDefaultValue", record.defaultStringField());
-  }
-
-  @Test
-  @DisplayName("Testing field with default value is set to provided value if not null")
-  public void testFieldWithDefaultValueIsSetToProvidedValueIfProvidedValueIsNotNull() {
-    ExampleRecordWithDefaultFields record = new ExampleRecordWithDefaultFields(
-        "ignore", "someValue");
-    Assertions.assertEquals("someValue", record.defaultStringField());
-  }
-}
