+++ src/test/java/com/chrimle/example/standard/TestExampleRecordWithBooleanFields.java
@@ -1,0 +1,31 @@
+package com.chrimle.example.standard;
+
+import com.chrimle.example.utils.AssertionUtils;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+@DisplayName("Testing Record with Boolean fields")
+public class TestExampleRecordWithBooleanFields {
+
+  private final Class<?> classUnderTest = ExampleRecordWithBooleanFields.class;
+
+  @Test
+  @DisplayName("Testing class generated is a Record")
+  public void testClassGeneratedIsRecord() {
+    AssertionUtils.assertIsRecord(classUnderTest);
+  }
+
+  @Test
+  @DisplayName("Testing generated Record is not annotated as @Deprecated")
+  public void testRecordIsNotAnnotatedAsDeprecated() {
+    AssertionUtils.assertClassIsNotAnnotatedAsDeprecated(classUnderTest);
+  }
+
+  @Test
+  @DisplayName("Testing generated Record has expected number of fields of expected types")
+  public void testRecordHasNumberOfFieldsOfType() {
+    AssertionUtils.assertRecordHasNumberOfFieldsOfType(
+        classUnderTest, 3, Boolean.class);
+  }
+
+}
